Ajout du controller+service+repository+model du Vehicule

+++ src/main/java/com/vroom/vroom/model/User.java
@@ -27,7 +27,7 @@ public class User {
         this.isAdmin = isAdmin;
     }
 
-    public Long getIdUser() {
+    public int getIdUser() {
         return idUser;
     }
